pridany formular na registraciu aj s kontrolnymi hlaskami login pri zlom zadani upozorni uzivatela pri logine sa zapise cas prihlasenia

+++ src/main/java/sk/upjs/drivingSchool/login/UserSession.java
@@ -37,7 +37,7 @@ public class UserSession {
         return userId;
     }
 
-    public void setUserId(final int userId) {
+    public void setUserId(final long userId) {
         this.userId = userId;
     }
